#210803_백정연_플래너 페이지 버튼 css 수정 및 controller 생성 date : 210803 writer : 백정연 branch : jeong comment : - 플래너 목록 : 삭제 버튼 텍스트 위치 수정 - 용도별로 분류 후 controller 생성 및 메소드명, mapping 주소 변경
