+++ src/main/resources/static/_css/mainStyle.css
@@ -2589,6 +2589,7 @@ table {
     border-radius: 7px;
     font-family: 'Jua', sans-serif;
     font-weight: bolder;
+    line-height:25px;
 }
 
 .planL-popupcontainer {
